10477 Move RateBands reads into companion obj
