+++ app/uk/gov/hmrc/tai/model/domain/RateBand.scala
@@ -16,4 +16,22 @@
 
 package uk.gov.hmrc.tai.model.domain
 
+import play.api.libs.json.{JsArray, JsValue}
+
 case class RateBand(income: BigDecimal, rate: BigDecimal)
+
+object RateBand {
+  def incomeAndRateBands(json: JsValue): Seq[RateBand] = {
+
+    val bands = (json \ "totalLiabilityDetails" \ "nonSavings" \ "taxBandDetails").asOpt[JsArray]
+    val details = bands.map(_.value.collect {
+      case js if (js \ "income").asOpt[BigDecimal].isDefined =>
+        RateBand((js \ "income").as[BigDecimal], (js \ "rate").as[BigDecimal])
+    })
+
+    details match {
+      case Some(rateBands) => rateBands.toSeq.sortBy(-_.rate)
+      case None            => Seq.empty[RateBand]
+    }
+  }
+}
